Refactoring URLs to be app-contained
Seeing as we're mostly one app -- with the potential for pluggable ones
like Norlønn, or a Visma eAccounting integration -- it is reasonable
that all the URLs belonging to us actually are contained within us as
well. After all, the projects URL mappings should be separated from our
apps available endpoints.