Added code coverage
Including branch coverage, a cute little badge, but also setup.cfg where
the coverage settings reside.

Note that you need to configure the COVERALLS_TOKEN in wercker for it to
get it, but you can also do this locally with something like
  X_COVERALLS_TOKEN=token-token-token wercker build

Works fairly well. Let's try to get the coverage up.