Society to societies data migration and fields
Relatively big change. We remove the society OneToOneField, and replace
it with a ManyToManyField supported by a Employment "via" class.

This means that in addition to querying costing a bit more, although
c'mon, we're not big, we have to create employment relations between
workers to facilitate who they work for.

On the plus side we can now have workers work for many different
societies, which is really convenienent for the few that do, and for us
wanting a single unified employment relationship.
After all, there's no reason a different individual should in any way be
duplicated just because of an oversight on our part in terms of use
cases.