Performance improvements for events
Using both select_related() and more, including annotating the default
queries with total hours and sum of shifts for events.

We almost always want it, so I think we're alright to do it by default.
Might create another manager if I'd like to keep it smooth, seeing as
there's Event.shifts as well.