Add initial poor tests and new set_up_user
Note that the set_up_user is meant to replace set_up_superuser, and is
currently quite filthy. WIP.

The tests for Norlønn shouldn't even be here; the wage reports can be
merged into the main application, while converting the Norlønn-module
into a output-only module, as that is the only direct supplier-related
functionality.