fix(tests): debug toolbar works under testing
Uses settings.DEBUG as the test setup will change settings.DEBUG to
False, whereas our DEBUG in settings.py will not by default do that.

Furthermore, always adds the debug URLs to the application's URL
patterns, resolving the KeyError that could occur when djdt is partially
loaded, but not fully, when running tests with SPBM_DEBUG set to True in
the parent environment.